refactor(app): extract message opening from inputs
